refactor: extract batch input helpers and operation validators into dedicated modules
Move OpInput helpers, normalizers, and OPERATION_VALIDATORS out of batch.ts into batchInputHelpers.ts and batchOperationValidators.ts to reduce route file size and improve testability.
